Move the `Book` class in its own source file.
